requirements.txt
# =============================
# Bot Telegram & Async
# =============================
aiogram==2.25.1
aiohttp==3.8.5
multidict==6.0.4
yarl==1.9.2
async-timeout==4.0.2

# =============================
# Banco de dados
# =============================
SQLAlchemy==2.0.21
psycopg2-binary==2.9.9
alembic==1.12.0

# =============================
# Cache / Filas / Redis
# =============================
redis==5.1.0

# =============================
# HTTP / Requests / APIs
# =============================
requests==2.31.0
python-dotenv==1.0.1

# =============================
# Web server / Worker
# =============================
gunicorn==23.0.0
uvicorn==0.23.2
fastapi==0.101.1

# =============================
# Segurança / Criptografia
# =============================
cryptography==41.0.3
python-jose==3.3.0

# =============================
# Observabilidade / Métricas
# =============================
prometheus-client==0.17.1

# =============================
# Extras / Typing
# =============================
typing-extensions==4.7.1